Adding docs and JSON column support
